+++ src/main/java/iris/Ui.java
@@ -1,3 +1,5 @@
+package src.main.java;
+
 import java.util.Scanner;
 
 public class Ui {
