+++ src/main/java/iris/command/DeleteTaskCommand.java
@@ -1,21 +1,22 @@
-public class MarkTaskCommand extends Command {
-    int index;
+package src.main.java.command;
 
-    public MarkTaskCommand(int i) {
-        this.index = i;
+public class DeleteTaskCommand extends Command {
+    int i;
+    public DeleteTaskCommand(int i) {
+        this.i = i;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, TaskStore taskStore) throws IrisException {
         Task task;
         try {
-            task = tasks.get(index);
+            task = tasks.get(i);
+            tasks.remove(i);
         } catch (IndexOutOfBoundsException e) {
             throw new UnknownTaskException();
         }
-        task.mark();
         taskStore.updateTasks(tasks);
-        Ui.output("Good job! I've marked this task done:");
+        Ui.output("I've removed this task");
         Ui.output(task.toString());
     }
 }
