+++ src/main/java/iris/command/FilterCommand.java
@@ -1,3 +1,5 @@
+package src.main.java.command;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
