+++ src/main/java/iris/command/ListCommand.java
@@ -1,3 +1,5 @@
+package src.main.java.command;
+
 public class ListCommand extends Command{
     @Override
     public void execute(TaskList tasks, Ui ui, TaskStore taskStore) {
