+++ src/main/java/iris/exception/MissingFieldException.java
@@ -1,3 +1,5 @@
+package src.main.java.exception;
+
 public class MissingFieldException extends IrisException {
     public MissingFieldException(String field) {
         super(field + " cannot be empty.");
