+++ src/main/java/iris/exception/NoTaskException.java
@@ -1,3 +1,5 @@
+package src.main.java.exception;
+
 public class NoTaskException extends IrisException {
     public NoTaskException() {
         super("I'm sorry, I don't know what you mean.");
