+++ src/main/java/iris/task/Deadline.java
@@ -1,3 +1,5 @@
+package src.main.java.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
