+++ src/main/java/iris/task/Event.java
@@ -1,3 +1,5 @@
+package src.main.java.command;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
