+++ src/main/java/iris/task/Todo.java
@@ -1,3 +1,5 @@
+package src.main.java.task;
+
 public class Todo extends Task{
     Todo(String name) {
         super(name);
